KAB-998 feat: annotate tools with session context, split workspace manager and sql tools due to circular imports, update tests
