+++ src/keboola_mcp_server/tools/storage.py
@@ -10,8 +10,8 @@ from pydantic import AliasChoices, BaseModel, Field, model_validator
 from keboola_mcp_server.client import KeboolaClient
 from keboola_mcp_server.config import MetadataField
 from keboola_mcp_server.errors import tool_errors
-from keboola_mcp_server.mcp import KeboolaMcpServer
-from keboola_mcp_server.tools.sql import WorkspaceManager
+from keboola_mcp_server.mcp import KeboolaMcpServer, with_session_state
+from keboola_mcp_server.tools.workspace import WorkspaceManager
 
 LOG = logging.getLogger(__name__)
 
@@ -140,6 +140,7 @@ class UpdateDescriptionResponse(BaseModel):
 
 
 @tool_errors()
+@with_session_state()
 async def get_bucket_detail(
     bucket_id: Annotated[str, Field(description='Unique ID of the bucket.')], ctx: Context
 ) -> BucketDetail:
@@ -152,6 +153,7 @@ async def get_bucket_detail(
 
 
 @tool_errors()
+@with_session_state()
 async def retrieve_buckets(ctx: Context) -> list[BucketDetail]:
     """Retrieves information about all buckets in the project."""
     client = KeboolaClient.from_state(ctx.session.state)
@@ -162,6 +164,7 @@ async def retrieve_buckets(ctx: Context) -> list[BucketDetail]:
 
 
 @tool_errors()
+@with_session_state()
 async def get_table_detail(
     table_id: Annotated[str, Field(description='Unique ID of the table.')], ctx: Context
 ) -> TableDetail:
@@ -187,6 +190,7 @@ async def get_table_detail(
 
 
 @tool_errors()
+@with_session_state()
 async def retrieve_bucket_tables(
     bucket_id: Annotated[str, Field(description='Unique ID of the bucket.')], ctx: Context
 ) -> list[TableDetail]:
@@ -203,6 +207,7 @@ async def retrieve_bucket_tables(
 
 
 @tool_errors()
+@with_session_state()
 async def update_bucket_description(
     bucket_id: Annotated[str, Field(description='The ID of the bucket to update.')],
     description: Annotated[str, Field(description='The new description for the bucket.')],
@@ -226,6 +231,7 @@ async def update_bucket_description(
 
 
 @tool_errors()
+@with_session_state()
 async def update_table_description(
     table_id: Annotated[str, Field(description='The ID of the table to update.')],
     description: Annotated[str, Field(description='The new description for the table.')],
@@ -251,6 +257,7 @@ async def update_table_description(
 
 
 @tool_errors()
+@with_session_state()
 async def update_column_description(
     table_id: Annotated[str, Field(description='The ID of the table that contains the column.')],
     column_name: Annotated[str, Field(description='The name of the column to update.')],
