+++ tests/tools/test_storage.py
@@ -7,7 +7,6 @@ from pytest_mock import MockerFixture
 
 from keboola_mcp_server.client import KeboolaClient
 from keboola_mcp_server.config import Config, MetadataField
-from keboola_mcp_server.tools.sql import TableFqn, WorkspaceManager
 from keboola_mcp_server.tools.storage import (
     BucketDetail,
     TableColumnInfo,
@@ -21,6 +20,7 @@ from keboola_mcp_server.tools.storage import (
     update_column_description,
     update_table_description,
 )
+from keboola_mcp_server.tools.workspace import TableFqn, WorkspaceManager
 
 
 def parse_iso_timestamp(ts: str) -> datetime:
